Change the heuristic from "is metadata locally availble" to "is it cheap to get metadata"
The rationale behind this change is that answering the question "do you have it locally" can in itself be very expensive. It also causes problems
with repository chains, because it's not because one of them has it locally, that we're not going to hit the network (because order of repositories
matter).

This commit changes the question from "is it available locally" to "is is cheap to get the answer"? If it's cheap to get the answer, then we're going
to perform the request serially. If it's not, we're putting it in the queue of work to be done concurrently. When a repository chain is hit, we then
make the difference between the 3 cases:

- the repository has the metadata at hand, and it's authoritative (it's available locally and up-to-date)
- the repository doesn't have the metadata at hand, but it's relatively cheap to get it, or tell that it's missing
- the repository doesn't have the metadata at hand, and it's expensive to get it or tell it's missing

Then, we can tell that a chain is cheap if, and only if, one of the repositories can get the authoritative answer fast, before those who are expensive,
or, all repositories are cheap.