Index services by type
This commit optimizes service lookup, by creating an index by type. When we're looking for a service of a specific type,
instead of iterating over all possible providers, we're now going to search in constant time. The disadvantage of this
technique is that registering services take more time (we need to index all interfaces and superclasses) and memory.